requirements.txt
openai
python-dotenv
reportlab
